feat(git): Improve performance when parsing git commits
The current approach iterates over all commits, even if these are not relevant for the desired node/path.
When using the oxidized’s web view, the sub path /nodes/version?node_full=[…] takes clearly over 30 seconds when having a commit count of 37000+. With this change, that time drops to not even one second, making the application actually usable and not nerv-ewreckingly slow.

Notes:
- We 'alias' the imported Git class to avoid name clashes
- We instanciate the log class separately, with false as 2nd parameter, to bypass 'count = 30' on the initialize method. Reason: All commits (=version) shall be displayed